+++ R/h_map_for_count_abnormal.R
@@ -23,7 +23,7 @@
 #' adlb <- synthetic_cdisc_data("latest")$adlb
 #' adlb <- df_explicit_na(adlb)
 #'
-#' tern::h_map_for_count_abnormal(
+#' tern:::h_map_for_count_abnormal(
 #'   df = adlb,
 #'   variables = list(anl = "ANRIND", split_rows = c("LBCAT", "PARAM")),
 #'   abnormal = list(low = c("LOW"), high = c("HIGH")),
@@ -51,7 +51,7 @@
 #'   ANRHI = rep(20, 12)
 #' )
 #' df$ANRIND <- factor(df$ANRIND, levels = c("LOW", "HIGH", "NORMAL"))
-#' tern::h_map_for_count_abnormal(
+#' tern:::h_map_for_count_abnormal(
 #'   df = df,
 #'   variables = list(
 #'     anl = "ANRIND",